Move common components inside folder like Wallet
